Moved files for login, main and register
Moved login and register, renamed main to constants
