+++ js/url/constants.mjs
@@ -1,8 +1,11 @@
 export const api_Url_Base = "https://api.noroff.dev/api/v1/social";
 export const api_register = `${api_Url_Base}/auth/register`;
 export const api_login = `${api_Url_Base}/auth/login`;
+export const api_posts = `${api_Url_Base}/posts`;
 
 export const methodPost = "post";
+export const methodPut = "put";
+export const methodDelete = "delete";
 export const pathProfile = "../profile.html";
 
 // export const POST_Options = {
