+++ signup.html
@@ -169,6 +169,6 @@
       </div>
     </footer>
 
-    <script type="module" src="js/register.mjs"></script>
+    <script type="module" src="js/formsregister.mjs"></script>
   </body>
 </html>
